ionic: remove support for event queues
The code has been unused for years, and never went into the
upstream driver.  It was originally to help support "scale"
model for macvlan offload.  The macvlan offload feature was
removed from the driver long ago,
    Remove macvlan offload (#27488)
and this code should have been yanked out at the same time as
it only adds complexity and confusion.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>