ionic: Reduce race conditions in ionic_device_reset()
Currently there is a chance that the qcqs could be NULL when the
following call path happens:

ionic_lif_deferred_work()
  -> ionic_link_status_check()
    -> ionic_start_queues()

This can happen if an operation is triggered that causes the queues to
be re-configured via the deferred work task and a ionic_device_reset
is called directly after. This is problematic because
ionic_device_reset() frees the qcqs, which then may be refenced in the
deferred work task thread causing a NULL pointer de-reference.

Fix this by stopping/flushing the watchdog timer and deferred work
at the beginning of ionic_device_reset() and then restarting the
watchdog timer at the end of ionic_device_reset().

As long as we aren't already running deferred_work on the fw_down
portion of the IONIC_DW_TYPE_LIF_RESET, there seems to be no other way
the watchdog timer handler will get restarted.  Also, since we detach
the netdev and are under the rtnl_lock() for ethtool, we don't have to
worry about the deferred_work being re-scheduled by anything but the
watchdog timer handler, which will only happen in the above case.
So, it seems like the only case we could get hosed is if we get an
IONIC_EVENT_RESET from FW. Fortunately an IONIC_EVENT_RESET isn't
received when triggering a reset from the driver.

Signed-off-by: Brett Creeley <brett.creeley@amd.com>