ionic_mnic: selectable DMA mask for giglio use
Giglio has a smaller memory map, so set up a module parameter to
allow for selecting a smaller DMA mask size.  It defaults to the
existing 52 bits, but can be set on the insmod command line to a
different size.  When the driver is running, the current size can
be inspected by reading /sys/module/ionic/parameters/asic_addr_len

Example use:
    # insmod ionic_mnic.ko asic_addr_len=42

    # cat /sys/module/ionic/parameters/asic_addr_len
    42

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>