ionic: Don't overwrite the cyclecounter bitmask
The driver was incorrectly overwriting the cyclecounter bitmask, which
was truncating it and not aligning to the hardware mask value. This
isn't causing any issues, but it's wrong. Fix this by not constraining
the cyclecounter/hardware mask.

Luckily, this seems to cause no issues, which is why this change
doesn't have a fixes tag and isn't being sent to net. However, if any
transformations from time->cycles are needed in the future, this
change will be needed.

Suggested-by: Allen Hubbe <allen.hubbe@amd.com>
Signed-off-by: Brett Creeley <brett.creeley@amd.com>