dsc-drivers: update drivers to 24.04.2-001
This is a driver update for testing of a different algorithm for handling
the missed-doorbell fix in the ionic driver.

A latency test in a scaled out setting (many VMs with many queues)
uncovered an issue with our earlier missed doorbell fix.  The basic
issue is that we're delaying things with a timer for every single queue
periodically using mod_timer() to reset itself, and mod_timer() becomes
a more and more expensive thing as there are more and more VFs and queues
each with their own timer.  A ping-pong latency test tends to exacerbate
the effect such that every napi is doing a mod_timer() in every cycle.

This version replaces all the per-queue timers with a new workqueue per
device, using a periodically workitem scheduled on the AdminQ's last
cpu which queues a workitem per running queue, each of which will call
napi_schedule on the specific queue if it hasn't already been run within
a certain time period.

This update is intended only for the PCI host driver use.  There are
a couple other changes related to some rework in our internal platform
support - those can be ignored for the PCI host driver.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>