pds: pull pds drivers branch into master
* pds: add the PDS host drivers

Add the current 1.58.0-53 snapshot of the PDS host drivers:
pds_core, pds_vfio, and pds_vdpa.  These drivers are host
drivers and not intended to be compiled for or used on the
DSC firmware, only on a host server with a DSC installed.

These drivers are to be built against Linux v6.1 or greater.

Slightly stripped-down versions of these have recently been
published to the upstream Linux mailing lists and are under
discussion.

Link: https://lore.kernel.org/netdev/20221207004443.33779-1-shannon.nelson@amd.com/
Link: https://lore.kernel.org/netdev/20221207010705.35128-1-brett.creeley@amd.com/
Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds: update the initial draft code to current day

A lot has happened with the PDS core and auxiliary modules after the
initial snapshot was taken.  Most of the history can be found in the
upstream Linux tree.  This update brings the out-of-tree code up to
the same level as what is currently upstream in Linux v6.12.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds_core: add support for fwctl auxiliary_device

This adds a new auxiliary_device to support the new fwctl
framework.  This new device is specific to each pds_core PF,
and there is only one pds_core PF per DSC, so there is one
connection for each DSC in the host.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds_fwctl: add pds_fwctl auxiliary_bus device driver

This adds support for the new fwctl framework for allowing a
userspace configuration program talk to the firmware.  It works
through an auxiliary_bus connection to the pds_core PF, and
there is only one PF per DSC, so there is one connection for
each DSC in the host.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds: update pds_core and pds_fwctl

Update the code to match the internal 1.105-a-15 FW source

Changes:
   - implement the RPC feature
   - remove the devlink ENABLE_FWCTL parametero

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds_core: fix dev info fw version handling

Address a compatibility issue between older and newer firmware
where the firmware module version list can be longer with the newer
firmware than originally expected.

Internal patches:
    pds_core: Core dev list info fix

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

* pds: add fwctl scope checking and code cleanup

Updates to match 1.105-a-42
 - Add the required fwctl operation scope checking to the pds_fwctl driver.
 - Fix an admiq overflow isue found in pds_core.

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>

---------

Signed-off-by: Shannon Nelson <shannon.nelson@amd.com>